+++ WebStore/Views/Shared/Partial/_MainMenu.cshtml
@@ -19,4 +19,6 @@
     </li>
     <li><a href="404.html">404</a></li>
     <li><a href="contact-us.html">Contact</a></li>
+    <li><a asp-controller="Employees" asp-action="Index">Работники</a></li>
+    
 </ul>
